shifted tests to 2tabular.py
