+++ test/test_3custom.py
@@ -2,7 +2,7 @@
 These were designed selfishly for my own uses.
 So seperating out, to one day sunset and remove.
 """
-from test.config import testing_parameters, testingtablename, LOCAL_FILE
+from test.config import testing_parameters, testingtablename
 import pytest
 
 
@@ -44,21 +44,3 @@ def test_revert_table(model):
         )
         == 1
     )
-
-
-@pytest.mark.parametrize("model", testing_parameters)
-def test_get_dependencies(model):
-    if len(model.Measures) > 0:
-        dependencies = model.Measures[0].get_dependencies()
-        assert len(dependencies) > 0
-    else:
-        assert True
-
-
-@pytest.mark.parametrize("model", testing_parameters)
-def test_get_sample_values(model):
-    if LOCAL_FILE[0] == "AdventureWorks Sales":
-        df = model.Columns["Country"].get_sample_values()
-        assert len(df) > 0
-    else:
-        assert True
